refactor(planning): extract the mobile by-goal row components (#467)
First slice of the planning-views split. Move MobilePlanningView's by-goal row
components — GoalAllocationRow, GoalItemRow, PlanLineItem (the fixed/insurance/
other line item) and the shared kebab MenuItem — into a new `planningRows.tsx`,
and lift the shared inline EditIcon / TrashIcon SVGs into `planningIcons.tsx`
(used by both the view and the rows). The components are already module-level and
purely prop-driven, so this is a relocation: the view still owns all data + action
handlers and passes them down. No behavior or markup change.

Now-unused lucide imports (MoreHorizontal / Check / X / TrendingUp) and
useCloseOnScroll drop out of the view.

Parity: the PlanningClient / MobilePlanningView component tests + the planning E2E
(create plan, month navigation — which render the goal rows) all green.

Verified: full unit suite 1261 passed; typecheck 0 errors; lint 0 errors;
E2E planning (3) passed. MobilePlanningView 1564 → 1216 lines.
